continued to organize and adapt the fighter creation function. Created utils in the fighter app to handle ad hoc scripting relevant to fighters and their associated objects.
